scanning_ze_magic_api/requirements.txt
click==8.1.3
Flask==2.2.3
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
numpy==1.24.2
opencv-python==4.7.0.72
packaging==23.0
Pillow==9.5.0
pytesseract==0.3.1
Werkzeug==2.2.3
